inventory, order & product finished

+++ src/product/product.service.ts
@@ -1,14 +1,27 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, Inject, forwardRef } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
+import { InventoryService } from 'src/inventory/inventory.service';
 
 @Injectable()
 export class ProductService {
-  constructor(private prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService, @Inject(forwardRef(() => InventoryService))
+    private readonly inventoryService: InventoryService,) {}
 
   async create(dto: CreateProductDto) {
-    return this.prisma.product.create({ data: dto });
+    const product = await this.prisma.product.create({ data: dto });
+    if(product){
+      await this.inventoryService.createMovement({
+        productId: product.id,
+        type: 'IN',
+        quantity: product.stock,
+        unitCost: product.cost,
+      });
+    }
+
+    return product;
+
   }
 
  async findAll(
@@ -74,7 +87,9 @@ export class ProductService {
   }
 
   async update(id: string, dto: UpdateProductDto) {
-    await this.findOne(id);
+    const product = await this.findOne(id);
+
+    if (!product) throw new NotFoundException('Producto no encontrado');
     
     return this.prisma.product.update({
       where: { id },
@@ -83,7 +98,9 @@ export class ProductService {
   }
 
   async remove(id: string) {
-    await this.findOne(id);
+    const product = await this.findOne(id);
+
+    if (!product) throw new NotFoundException('Producto no encontrado');
 
     return this.prisma.product.delete({ where: { id } });
   }
